Makefile

+++ src/my_scanf_storagefuncs.c
@@ -6,7 +6,7 @@
 /*   By: raica-ba <raica-ba@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/10 23:44:48 by raica-ba          #+#    #+#             */
-/*   Updated: 2025/03/15 22:55:03 by raica-ba         ###   ########.fr       */
+/*   Updated: 2025/03/15 23:26:08 by raica-ba         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ void	store_in_int(int *nb, int *j, char *buff)
 	*nb = ft_atoi(value);
 }
 
-void	store_in_float(float *dec_nb, int *j, char *buff)
+void	store_in_float(double *dec_nb, int *j, char *buff)
 {
 	char	value[50];
